add gant cgart with fake date

+++ front-end/vue/components/employee/load-employee.vue
@@ -58,26 +58,54 @@ export default {
          * delete item
          */
         deleteValue(data) {
+            const url = RouteHelper.routePath("api.project.task.data", {
+                projectId: "*",
+                employeeId: data.name,
+            });
+
+            return new Promise((resolve, reject) => {
+                LoadTaskHelper.loadTasks(url)
+                    .then((res) => {
+                        if (res.length <= 0) {
+                            const confirmed = confirm(
+                                `آیا از حذف این کارمند اطمینان دارید با کد ${data.name} ?`
+                            );
+                            if (confirmed) {
+                                const index = this.value.findIndex(
+                                    (x) => x.name == data.name
+                                );
+                                if (index > -1) {
+                                    const url = RouteHelper.routePath(
+                                        "api.employee.delete",
+                                        {
+                                            id: this.value[index]._id,
+                                        }
+                                    );
+
+                                    try {
+                                        LoadEmployeeHelper.deleteEmployee(url);
+                                        Vue.delete(this.value, index);
+                                    } catch (err) {
+                                        console.log(err);
+                                    }
+                                }
+                            }
+                        } else {
+                            alert(
+                                "پروژه انتخابی دارای وظیفه می باش  لطفا ابتدا وظیفه ها را حذف کنید"
+                            );
+                        }
+                    })
+                    .catch((err) => {
+                        reject(err);
+                    });
+            });
+
             const confirmed = confirm(
                 "Are you sure delete this employee with code : " +
                     data.name +
                     " ?"
             );
-            if (confirmed) {
-                const index = this.value.findIndex((x) => x.name == data.name);
-                if (index > -1) {
-                    const url = RouteHelper.routePath("api.employee.delete", {
-                        id: this.value[index]._id,
-                    });
-
-                    try {
-                        LoadEmployeeHelper.deleteEmployee(url);
-                        Vue.delete(this.value, index);
-                    } catch (err) {
-                        console.log(err);
-                    }
-                }
-            }
         },
         onChangeEmployee(payload) {
             const employee = payload.target.value;
